Move single-use v2 components from shared
Move them to relevant directories
