+++ src/components/shared/formItems/index.tsx
@@ -15,7 +15,6 @@ import ProjectReserved from './ProjectReserved'
 import ProjectTarget from './ProjectTarget'
 import ProjectTicketMods from './ProjectTicketMods'
 import ProjectTwitter from './ProjectTwitter'
-import ProjectReconfiguration from './ProjectReconfiguration'
 import TokenRefs from './TokenRefs'
 
 export const FormItems = {
@@ -34,7 +33,6 @@ export const FormItems = {
   ProjectPayoutMods,
   ProjectReserved,
   ProjectTarget,
-  ProjectReconfiguration,
   ProjectTicketMods,
   ProjectTwitter,
   TokenRefs,
