+++ src/components/shared/forms/RestrictedActionsForm.tsx
@@ -51,7 +51,7 @@ export default function RestrictedActionsForm({
           name="payIsPaused"
           label={t`Pause payments`}
           extra={t`Your project cannot receive direct payments while paused.`}
-          valuePropName="checked"
+          valuePropName={'checked'}
         >
           <Switch />
         </Form.Item>
